# ARROW-1942: [C++] Hash table specializations for small integers