Add hash table specializations for int8_t and uint8_t