@wesm 2018-02-09
@xuepanchen it looks like something went haywire in the merge (but it is OK, no need to change anything) -- in the future you can do `git pull --ff-only` or `git pull --rebase` when working on the branch