@xuepanchen 2018-02-09
@wesm got you. Compute-benchmark test failed when I run it locally and I can't figure out why. It seems to fail too before I added the 8-bit integer case