@jreback 2018-02-20
fyi in pandas we currently promote to int64 for smaller item size for hash operations