@wesm 2018-02-20
@jreback I thought so, thanks for confirming =)