@wesm 2018-02-20
@xuepanchen I added a template for the 8 bit hash function to avoid arithmetic in the uint8 case

before this change:

```
$ ./release/compute-benchmark --benchmark_filter=UInt8
Run on (8 X 4399.69 MHz CPU s)
2018-02-19 21:57:13
Benchmark                                                         Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_UniqueUInt8NoNulls/16M/200/min_time:1.000/real_time         8339 us       8339 us        166   1.87372GB/s
BM_UniqueUInt8WithNulls/16M/200/min_time:1.000/real_time      28536 us      28537 us         49     560.7MB/s
```

after this change:

```
$ ./release/compute-benchmark --benchmark_filter=UInt8
Run on (8 X 4400 MHz CPU s)
2018-02-19 21:55:51
Benchmark                                                         Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_UniqueUInt8NoNulls/16M/200/min_time:1.000/real_time         7749 us       7749 us        180   2.01641GB/s
BM_UniqueUInt8WithNulls/16M/200/min_time:1.000/real_time      28042 us      28042 us         50   570.571MB/s
```