@wesm 2018-02-20
Appveyor builds looking fine here: https://ci.appveyor.com/project/xuepanchen/arrow/build/1.0.25. Merging, thank you!